requirements.txt
google-api-python-client==2.99.0
googleapis-common-protos==1.59.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-telegram-bot==20.5
google-api-core==2.11.1
google-auth==2.22.0
oauth2client==4.1.3
schedule==1.2.0
pytube==15.0.0